company settings collapse added

+++ protected/controllers/ActController.php
@@ -255,8 +255,11 @@ class ActController extends Controller
                         $scope->save();
                     }
                 }
-                foreach ($oldScopes as $scope) {
-                    $scope->delete();
+                foreach ($oldScopes as $scopeId) {
+                    $scope = ActScope::model()->findByPk($scopeId);
+                    if ($scope) {
+                        $scope->delete();
+                    }
                 }
 
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : Yii::app()->createUrl('act/carwash'));
